requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.5
llama-parse>=0.1.1
pymupdf>=1.23.8
google-generativeai>=0.3.2
networkx>=3.2.1
rdflib>=7.0.0
deepeval>=0.20.9
python-dotenv>=1.0.0
tavily-python>=0.3.0
anthropic>=0.7.8
pandas>=2.1.4
numpy>=1.24.0
matplotlib>=3.8.2
python-dotenv>=1.0.0
langchain_anthropic>=0.0.10

